Move the canFeature policy out of the Activity vocabulary layer
Activity classes are generic vocabulary and must not contain application
logic. The canFeature computation (FEP-7aa9) moves from Activity\Actor
into a model-layer trait used by the User and Blog models. The generic
Actor class no longer injects the local site policy into remote actor
representations, and the Application actor stops advertising a consent
policy altogether.
