+++ .github/changelog/update-move-interaction-policy-to-models
@@ -1,0 +1,4 @@
+Significance: patch
+Type: fixed
+
+Publish the Starter Kit consent policy only on your own blog and author profiles, no longer on system or third-party profiles.
